backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
psycopg2-binary
python-dotenv
redis
rq
httpx
sentence-transformers
torch
rank-bm25
